fix: shared helpers
