+++ deployment/ccip/changeset/v1_5_1/cs_configure_token_pools_integ_test.go
@@ -32,22 +32,6 @@ import (
 	commonchangeset "github.com/smartcontractkit/chainlink/deployment/common/changeset"
 )
 
-// createSymmetricRateLimits is a utility to quickly create a rate limiter config with equal inbound and outbound values.
-func createSymmetricRateLimits(rate int64, capacity int64) v1_5_1.RateLimiterConfig {
-	return v1_5_1.RateLimiterConfig{
-		Inbound: token_pool.RateLimiterConfig{
-			IsEnabled: rate != 0 || capacity != 0,
-			Rate:      big.NewInt(rate),
-			Capacity:  big.NewInt(capacity),
-		},
-		Outbound: token_pool.RateLimiterConfig{
-			IsEnabled: rate != 0 || capacity != 0,
-			Rate:      big.NewInt(rate),
-			Capacity:  big.NewInt(capacity),
-		},
-	}
-}
-
 // validateMemberOfTokenPoolPair performs checks required to validate that a token pool is fully configured for cross-chain transfer.
 func validateMemberOfTokenPoolPair(
 	t *testing.T,
