+++ deployment/ccip/changeset/v1_5_1/cs_configure_token_pools_unit_test.go
@@ -3,9 +3,7 @@
 package v1_5_1_test
 
 import (
-	"bytes"
 	"math/big"
-	"sort"
 	"testing"
 	"time"
 
@@ -15,7 +13,6 @@ import (
 
 	"github.com/smartcontractkit/chainlink-ccip/chains/evm/gobindings/generated/v1_5_1/token_pool"
 	"github.com/smartcontractkit/chainlink-common/pkg/logger"
-	"github.com/smartcontractkit/chainlink-evm/gethwrappers/shared/generated/initial/burn_mint_erc677"
 
 	cldf_chain "github.com/smartcontractkit/chainlink-deployments-framework/chain"
 	cldf "github.com/smartcontractkit/chainlink-deployments-framework/deployment"
@@ -30,82 +27,6 @@ import (
 	commonchangeset "github.com/smartcontractkit/chainlink/deployment/common/changeset"
 	"github.com/smartcontractkit/chainlink/deployment/common/proposalutils"
 )
-
-// createSymmetricRateLimits is a utility to quickly create a rate limiter config with equal inbound and outbound values.
-func createSymmetricRateLimits(rate int64, capacity int64) v1_5_1.RateLimiterConfig {
-	return v1_5_1.RateLimiterConfig{
-		Inbound: token_pool.RateLimiterConfig{
-			IsEnabled: rate != 0 || capacity != 0,
-			Rate:      big.NewInt(rate),
-			Capacity:  big.NewInt(capacity),
-		},
-		Outbound: token_pool.RateLimiterConfig{
-			IsEnabled: rate != 0 || capacity != 0,
-			Rate:      big.NewInt(rate),
-			Capacity:  big.NewInt(capacity),
-		},
-	}
-}
-
-// validateMemberOfTokenPoolPair performs checks required to validate that a token pool is fully configured for cross-chain transfer.
-func validateMemberOfTokenPoolPair(
-	t *testing.T,
-	state stateview.CCIPOnChainState,
-	tokenPool *token_pool.TokenPool,
-	expectedRemotePools []common.Address,
-	tokens map[uint64]*cldf.ContractDeploy[*burn_mint_erc677.BurnMintERC677],
-	tokenSymbol shared.TokenSymbol,
-	chainSelector uint64,
-	rate *big.Int,
-	capacity *big.Int,
-	expectedOwner common.Address,
-) {
-	// Verify that the owner is expected
-	owner, err := tokenPool.Owner(nil)
-	require.NoError(t, err)
-	require.Equal(t, expectedOwner, owner)
-
-	// Fetch the supported remote chains
-	supportedChains, err := tokenPool.GetSupportedChains(nil)
-	require.NoError(t, err)
-
-	// Verify that the rate limits and remote addresses are correct
-	for _, supportedChain := range supportedChains {
-		inboundConfig, err := tokenPool.GetCurrentInboundRateLimiterState(nil, supportedChain)
-		require.NoError(t, err)
-		require.True(t, inboundConfig.IsEnabled)
-		require.Equal(t, capacity, inboundConfig.Capacity)
-		require.Equal(t, rate, inboundConfig.Rate)
-
-		outboundConfig, err := tokenPool.GetCurrentOutboundRateLimiterState(nil, supportedChain)
-		require.NoError(t, err)
-		require.True(t, outboundConfig.IsEnabled)
-		require.Equal(t, capacity, outboundConfig.Capacity)
-		require.Equal(t, rate, outboundConfig.Rate)
-
-		remoteTokenAddress, err := tokenPool.GetRemoteToken(nil, supportedChain)
-		require.NoError(t, err)
-		require.Equal(t, common.LeftPadBytes(tokens[supportedChain].Address.Bytes(), 32), remoteTokenAddress)
-
-		remotePoolAddresses, err := tokenPool.GetRemotePools(nil, supportedChain)
-		require.NoError(t, err)
-
-		require.Len(t, remotePoolAddresses, len(expectedRemotePools))
-		expectedRemotePoolAddressesBytes := make([][]byte, len(expectedRemotePools))
-		for i, remotePool := range expectedRemotePools {
-			expectedRemotePoolAddressesBytes[i] = common.LeftPadBytes(remotePool.Bytes(), 32)
-		}
-		sort.Slice(expectedRemotePoolAddressesBytes, func(i, j int) bool {
-			return bytes.Compare(expectedRemotePoolAddressesBytes[i], expectedRemotePoolAddressesBytes[j]) < 0
-		})
-		sort.Slice(remotePoolAddresses, func(i, j int) bool {
-			return bytes.Compare(remotePoolAddresses[i], remotePoolAddresses[j]) < 0
-		})
-		for i := range expectedRemotePoolAddressesBytes {
-			require.Equal(t, expectedRemotePoolAddressesBytes[i], remotePoolAddresses[i])
-		}
-	}
-}
 
 func TestValidateRemoteChains(t *testing.T) {
 	t.Parallel()
@@ -344,60 +265,60 @@ func TestValidateConfigureTokenPoolContracts(t *testing.T) {
 		{
 			Msg: "Configure new pools on registry",
 			RegistrationPass: &regPass{
-				SelectorA2B: createSymmetricRateLimits(100, 1000),
-				SelectorB2A: createSymmetricRateLimits(100, 1000),
+				SelectorA2B: testhelpers.CreateSymmetricRateLimits(100, 1000),
+				SelectorB2A: testhelpers.CreateSymmetricRateLimits(100, 1000),
 			},
 		},
 		{
 			Msg: "Configure new pools on registry, update their rate limits",
 			RegistrationPass: &regPass{
-				SelectorA2B: createSymmetricRateLimits(100, 1000),
-				SelectorB2A: createSymmetricRateLimits(100, 1000),
+				SelectorA2B: testhelpers.CreateSymmetricRateLimits(100, 1000),
+				SelectorB2A: testhelpers.CreateSymmetricRateLimits(100, 1000),
 			},
 			UpdatePass: &updatePass{
 				UpdatePoolOnA: false,
 				UpdatePoolOnB: false,
-				SelectorA2B:   createSymmetricRateLimits(200, 2000),
-				SelectorB2A:   createSymmetricRateLimits(200, 2000),
+				SelectorA2B:   testhelpers.CreateSymmetricRateLimits(200, 2000),
+				SelectorB2A:   testhelpers.CreateSymmetricRateLimits(200, 2000),
 			},
 		},
 		{
 			Msg: "Configure new pools on registry, update both pools",
 			RegistrationPass: &regPass{
-				SelectorA2B: createSymmetricRateLimits(100, 1000),
-				SelectorB2A: createSymmetricRateLimits(100, 1000),
+				SelectorA2B: testhelpers.CreateSymmetricRateLimits(100, 1000),
+				SelectorB2A: testhelpers.CreateSymmetricRateLimits(100, 1000),
 			},
 			UpdatePass: &updatePass{
 				UpdatePoolOnA: true,
 				UpdatePoolOnB: true,
-				SelectorA2B:   createSymmetricRateLimits(100, 1000),
-				SelectorB2A:   createSymmetricRateLimits(100, 1000),
+				SelectorA2B:   testhelpers.CreateSymmetricRateLimits(100, 1000),
+				SelectorB2A:   testhelpers.CreateSymmetricRateLimits(100, 1000),
 			},
 		},
 		{
 			Msg: "Configure new pools on registry, update only one pool",
 			RegistrationPass: &regPass{
-				SelectorA2B: createSymmetricRateLimits(100, 1000),
-				SelectorB2A: createSymmetricRateLimits(100, 1000),
+				SelectorA2B: testhelpers.CreateSymmetricRateLimits(100, 1000),
+				SelectorB2A: testhelpers.CreateSymmetricRateLimits(100, 1000),
 			},
 			UpdatePass: &updatePass{
 				UpdatePoolOnA: false,
 				UpdatePoolOnB: true,
-				SelectorA2B:   createSymmetricRateLimits(200, 2000),
-				SelectorB2A:   createSymmetricRateLimits(200, 2000),
+				SelectorA2B:   testhelpers.CreateSymmetricRateLimits(200, 2000),
+				SelectorB2A:   testhelpers.CreateSymmetricRateLimits(200, 2000),
 			},
 		},
 		{
 			Msg: "Configure new pools on registry with multiple token CS",
 			RegistrationPass: &regPass{
-				SelectorA2B: createSymmetricRateLimits(100, 1000),
-				SelectorB2A: createSymmetricRateLimits(100, 1000),
+				SelectorA2B: testhelpers.CreateSymmetricRateLimits(100, 1000),
+				SelectorB2A: testhelpers.CreateSymmetricRateLimits(100, 1000),
 			},
 			UpdatePass: &updatePass{
 				UpdatePoolOnA: false,
 				UpdatePoolOnB: true,
-				SelectorA2B:   createSymmetricRateLimits(200, 2000),
-				SelectorB2A:   createSymmetricRateLimits(200, 2000),
+				SelectorA2B:   testhelpers.CreateSymmetricRateLimits(200, 2000),
+				SelectorB2A:   testhelpers.CreateSymmetricRateLimits(200, 2000),
 			},
 			runWithMultipleTokenCS: true,
 		},
